change init > scan
